Added filter language feature Added filter user feature Clean up code in clear object and commands object Added scraping data off Reddit Added JRAW maven dependency
TODO: COMMENT
TODO: Clean up messy codes
TODO: Fix any bug
TODO: Fix hardcoded values
TODO: Fix all 12 warnings

+++ src/main/java/Commands/Clear.java
@@ -5,43 +5,39 @@ import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import Main.main;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
+import static net.dv8tion.jda.api.Permission.ADMINISTRATOR;
+
 public class Clear extends ListenerAdapter {
 
-    public void onGuildMessageReceived(GuildMessageReceivedEvent event){
+    public void onGuildMessageReceived(@NotNull GuildMessageReceivedEvent event) {
         String[] args = event.getMessage().getContentRaw().split("\\s+");
 
-        if(args[0].equalsIgnoreCase(main.prefix + "clear")){
-            if(args.length < 2){
+        if (args[0].equalsIgnoreCase(main.prefix + "clear") && event.getMessage().getMember().hasPermission(ADMINISTRATOR)) {
+            if (args.length < 2) {
                 // Error
-                System.out.println("A");
                 event.getChannel().sendTyping().queue();
                 event.getChannel().sendMessage("Need more than one argument").queue();
-            }else{
-                try{
+            } else {
+                try {
                     List<Message> messages = event.getChannel().getHistory().retrievePast(Integer.parseInt(args[1])).complete();
                     event.getChannel().deleteMessages(messages).queue();
-                    event.getChannel().sendMessage("Successfully deleted " + Integer.parseInt(args[1])+ " Message").queue();
-                }catch(IllegalArgumentException e){
-                    if(e.toString().startsWith("java.lang.IllegalArgumentException: Message retrieval")){
-                        EmbedBuilder error = new EmbedBuilder();
-                        error.setColor(0xff3923);
+                    event.getChannel().sendMessage("Successfully deleted " + Integer.parseInt(args[1]) + " Message").queue();
+                } catch (IllegalArgumentException e) {
+                    EmbedBuilder error = new EmbedBuilder();
+                    error.setColor(0xff3923);
+                    System.out.println(e.toString());
+                    if (e.toString().startsWith("java.lang.IllegalArgumentException: Must provide at least 2 or at most 100 messages to be deleted."))
                         error.setTitle("Too many / too little message selected");
-                        error.setDescription("1-100 messages");
-                        event.getChannel().sendTyping().queue();
-                        event.getChannel().sendMessage(error.build()).queue();
-                        error.clear();
-                    }else{
-                        EmbedBuilder error = new EmbedBuilder();
-                        error.setColor(0xff3923);
+                    else
                         error.setTitle("Trying to delete message older than 2 week");
-                        error.setDescription("1-100 messages");
-                        event.getChannel().sendTyping().queue();
-                        event.getChannel().sendMessage(error.build()).queue();
-                        error.clear();
-                    }
+                    error.setDescription("2-100 messages");
+                    event.getChannel().sendTyping().queue();
+                    event.getChannel().sendMessage(error.build()).queue();
+                    error.clear();
                 }
             }
         }
